[BACKEND] feature(exceptions): adicionar mensagens
