+++ backend/src/core/auth/login/login.strategy.ts
@@ -15,7 +15,7 @@ export class LoginPassportLocalStrategy extends PassportStrategy(Strategy) {
     const user = await this.authService.validateUser(email, password);
 
     if (!user) {
-      throw new UnauthorizedException();
+      throw new UnauthorizedException('E-mail ou senha incorretos.');
     }
 
     return user;
