+++ frontend/src/exceptions/HttpException.ts
@@ -1,0 +1,7 @@
+export class HttpException extends Error {
+  constructor(message: string) {
+    super(message)
+    this.name = 'HttpException'
+    Object.setPrototypeOf(this, HttpException.prototype)
+  }
+}
