--- frontend/src/exceptions/http-error.ts
@@ -1,8 +1,0 @@
-export class HttpError extends Error {
-    constructor(message: string) {
-      super(message);
-      this.name = "HttpError";
-      Object.setPrototypeOf(this, HttpError.prototype);
-    }
-  }
-
